Improve record_read(), fix bug and remove span_read_bytes()
This function was inconsistent with all other read functions.
It would read the data and then slice into the array, instead of reading
byte by byte. After fixing that, span_read_bytes() because unused.

Also fixed a bug where sector was being read as u32 instead of u16.
This worked because the next two bytes are 0, but was a bug nonetheless.